eve_industry_tool/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
httpx
jinja2
python-dotenv
python-multipart
itsdangerous
